+++ src/main/kotlin/com/example/mtservice/service/BalanceService.kt
@@ -1,6 +1,6 @@
 package com.example.mtservice.service
 
 interface BalanceService {
-    fun add(id: Long, amount: Double)
-    fun get(id: Long): Double
+    fun add(id: Long, amount: Long)
+    fun get(id: Long): Long
 }
