+++ src/main/kotlin/com/example/mtservice/service/ProfileService.kt
@@ -4,5 +4,5 @@ import com.example.mtservice.dto.BalanceDto
 
 interface ProfileService {
     fun getAllBalances(userId: Long): List<BalanceDto>
-    fun addNewBalance(userId: Long, amount: Double)
+    fun addNewBalance(userId: Long, amount: Long)
 }
